+++ CHANGELOG.rst
@@ -17,7 +17,7 @@ are used for versioning (schema follows below):
 
 0.2.12
 ------
-2021-11-25
+2021-12-06
 
 - Tested against Django 3.1 and 3.2.
 - Tested against Python 3.9 and 3.10.
@@ -25,7 +25,7 @@ are used for versioning (schema follows below):
 - Drop Python 2.x support.
 - Drop Python 3.5 support.
 - Drop Django < 2.2 support.
-- Drop Django REST Framework < 3.9.x support.
+- Drop Django REST Framework < 3.9 support.
 
 0.2.11
 ------